Use 4-wide f32 vectors, drop trig randomUnit
Switched Vec3 from @Vector(3, f64) to @Vector(4, f32) to actually use SIMD. Replaced randomUnit with rejection sampling, and seed the per thread PRNG deterministically so threads no longer share a single random stream. Dropped some dead code.

Saved about ~3 seconds on the render settings I used as benchark. Bit of a waste of time.

+++ src/hittable.zig
@@ -15,7 +15,7 @@ const Material = material.Material;
 pub const hitRecord = struct {
     p: Point3,
     normal: Vec3,
-    t: f64,
+    t: f32,
     frontFace: bool,
     mat: Material,
